requirements.txt
requests>=2.25.1
psutil>=5.8.0
python-dotenv>=0.19.0 